fix(next-prompt): suppress unbounded hidden samples
